Only set the parent and the resource when alert definition is created, not every time it's updated

+++ src/org/hyperic/hq/events/server/session/AlertDefinition.java
@@ -223,8 +223,8 @@ public class AlertDefinition
         _priority = priority;
     }
 
-    public int getAppdefId() {
-        return getResource().getInstanceId().intValue();
+    public Integer getAppdefId() {
+        return getResource().getInstanceId();
     }
 
     public int getAppdefType() {
